# Conflicts:
#	apollo-workspace/apollo/src/main/resources/templates/common/surveys/components/question_group/question_group_container.html
#	apollo-workspace/apollo/src/main/resources/templates/common/surveys/components/questions/multi_choice.html
#	apollo-workspace/apollo/src/main/resources/templates/common/surveys/components/questions/single_choice.html
#	apollo-workspace/apollo/src/main/resources/templates/common/surveys/components/questions/text_input.html

+++ apollo-workspace/apollo/src/main/java/it/univaq/disim/mwt/apollo/business/impl/SurveyServiceImpl.java
@@ -102,10 +102,13 @@ public class SurveyServiceImpl implements SurveyService {
 	@Override
 	public void deleteSurvey(Survey survey) throws BusinessException {
 		
-		Iterable<QuestionGroup> groups = (Iterable<QuestionGroup>)survey.getQuestionGroups();
-
-		questionGroupService.deleteQuestionGroupList(groups);
+		if (survey.getQuestionGroups() != null && survey.getQuestionGroups().size() > 0) {
+			Iterable<QuestionGroup> groups = (Iterable<QuestionGroup>)survey.getQuestionGroups();
+			questionGroupService.deleteQuestionGroupList(groups);
+		} 
+		
 		surveyRepository.delete(survey);
+
 	}
 
 }
